make modifiers for nav APIs and move to separate files
